fix(story): add ai-service for content generation
